+++ CHANGELOG.md
@@ -13,3 +13,4 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 - Adding function that always choose the best arm to pull
 - Adding class of estimators for mult-armed-bandits
 - implemented two ways to track means of arm rewards (stationary and non-stationary)
+- implemented a greedy aprouch to solve the mult-armed bandit stationary or not